+++ packages/synapse-interface/pages/state-managed-bridge/index.tsx
@@ -73,6 +73,7 @@ import {
 } from '@/components/Maintenance/Maintenance'
 import { wagmiConfig } from '@/wagmiConfig'
 import { useStaleQuoteUpdater } from '@/utils/hooks/useStaleQuoteUpdater'
+import { screenAddress } from '@/utils/screenAddress'
 
 const StateManagedBridge = () => {
   const { address } = useAccount()
@@ -349,6 +350,14 @@ const StateManagedBridge = () => {
 
   const executeBridge = async () => {
     let pendingPopup: any
+
+    if (destinationAddress) {
+      const isRisky = await screenAddress(destinationAddress)
+      if (isRisky) {
+        return
+      }
+    }
+
     segmentAnalyticsEvent(
       `[Bridge] initiates bridge`,
       {
@@ -362,6 +371,7 @@ const StateManagedBridge = () => {
         destinationToken: toToken?.routeSymbol,
         exchangeRate: BigInt(bridgeQuote.exchangeRate.toString()),
         routerAddress: bridgeQuote.routerAddress,
+        bridgeQuote,
       },
       true
     )
@@ -455,6 +465,7 @@ const StateManagedBridge = () => {
         destinationToken: toToken?.routeSymbol,
         exchangeRate: BigInt(bridgeQuote.exchangeRate.toString()),
         routerAddress: bridgeQuote.routerAddress,
+        bridgeQuote,
       })
       dispatch(
         updatePendingBridgeTransaction({