+++ services/rfq/relayer/relapi/model.go
@@ -24,3 +24,12 @@ type PutRelayAckResponse struct {
 	ShouldRelay    bool   `json:"should_relay"`
 	RelayerAddress string `json:"relayer_address"`
 }
+
+// GetQuoteRequestResponse is the response to a get quote request.
+type GetQuoteRequestResponse struct {
+	QuoteRequestRaw string `json:"quote_request"`
+	OriginChainID   uint32 `json:"origin_chain_id"`
+	DestChainID     uint32 `json:"dest_chain_id"`
+	OriginToken     string `json:"origin_token"`
+	DestToken       string `json:"dest_token"`
+}