+++ packages/contracts-rfq/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.11](https://github.com/synapsecns/sanguine/compare/FastBridge@0.2.10...FastBridge@0.2.11) (2024-07-16)
+
+**Note:** Version bump only for package FastBridge
+
+
+
+
+
+## [0.2.10](https://github.com/synapsecns/sanguine/compare/FastBridge@0.2.9...FastBridge@0.2.10) (2024-07-03)
+
+**Note:** Version bump only for package FastBridge
+
+
+
+
+
 ## [0.2.9](https://github.com/synapsecns/sanguine/compare/FastBridge@0.2.8...FastBridge@0.2.9) (2024-06-09)